Prepare APIRequest and Interface
The APIRequest is in a usable state. The approach would be:
'from APIRequest import <one of eligable requests>' for example.

    from APIRequest import TeamsRequest

    league = TeamsRequest.makeapicall()

and league would now be complete with all the team information.

This commit is another snapshot though of the current design state. Because from
here I think I am going to go back to use APIRequest and have each request as
classmethod as a factory method to create the OWL model classes.
